% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/da_homogenization.R
\name{.da_homogenization}
\alias{.da_homogenization}
\title{Homogenize Differential Abundance Results}
\usage{
.da_homogenization(da_result)
}
\arguments{
\item{da_result}{A tibble containing the differential abundance results.
The data frame should have columns corresponding to either MaAsLin2 or
DESeq2 output formats.}
}
\value{
A homogenized tibble with a consistent format regardless of the
original DA result format.
}
\description{
This internal function takes differential abundance analysis results in a
tibble format from either DESeq2 or MaAsLin2 and homogenizes them into a
consistent format. The function first determines the type of input
(DESeq2/MaAsLin2) based on the column names and then calls for another
internal function to create a new tibble with homogenized column names. If
the input type is unrecognized, an error is raised.
}
