% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/da_homogenization.R
\name{.deseq_homogenize}
\alias{.deseq_homogenize}
\title{Homogenize DESeq2 Differential Abundance Results}
\usage{
.deseq_homogenize(tbl)
}
\arguments{
\item{tbl}{A tibble containing the DESeq2 DA analysis results.}
}
\value{
A tibble with a homogenized format.
}
\description{
This internal function homogenizes DESeq2 differential abundance (DA)
results into a consistent format with specific column names. The function
processes the input tibble and ensures that the output tibble has a
consistent structure with columns relevant for downstream analysis. It fills
certain columns with placeholder values as they are not applicable to
DESeq2 results.
}
