% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_lollipop.R
\name{.deseq_lollipop}
\alias{.deseq_lollipop}
\title{Plot DESeq2 Lollipop Plot}
\usage{
.deseq_lollipop(da_result)
}
\arguments{
\item{da_result}{A tibble containing the results of the DESeq2 DA analysis.}
}
\value{
A ggplot2 object representing the lollipop plot.
}
\description{
This function generates a lollipop plot for DESeq2 differential abundance
analysis results.
}
