% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_volcano.R
\name{.deseq_volcano}
\alias{.deseq_volcano}
\title{Plot DESeq2 Volcano Plot}
\usage{
.deseq_volcano(da_result)
}
\arguments{
\item{da_result}{A tibble containing the results of the DESeq2 DA analysis.}
}
\value{
A ggplot2 object representing the volcano plot.
}
\description{
This function generates a volcano plot for DESeq2 differential abundance
analysis results.
}
