% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_humann.R
\name{.extract_norm_method}
\alias{.extract_norm_method}
\title{Extract HUMAnN3 output data normalization method}
\usage{
.extract_norm_method(tbl)
}
\arguments{
\item{tbl}{A tibble containing the data from the HUMAnN3 output file.}
}
\value{
A character string indicating the normalization method ('cpm' or 'rpk').
}
\description{
This function determines if abundance data for each sample is in RPK units
(reads per kilobase) or is normalized to CPM units (copies per million). It
also ensures consistency of units across all samples (columns).
}
\keyword{internal}
