% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_humann.R
\name{.humann_metadata_check}
\alias{.humann_metadata_check}
\title{Check metadata consistency with HUMAnN3 output file}
\usage{
.humann_metadata_check(meta_df, humann_tbl)
}
\arguments{
\item{meta_df}{A tibble containing sample metadata.}

\item{humann_tbl}{A tibble containing the data from the HUMAnN3 output file.}
}
\value{
The input metadata tibble if all checks pass.
}
\description{
This function verifies the consistency of the provided metadata with the
HUMAnN3 file. It ensures metadata contains sufficient annotations and that
sample IDs in the metadata match the column names of the HUMAnN3 file.
}
\keyword{internal}
