% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_humann.R
\name{.humann_path_checks}
\alias{.humann_path_checks}
\title{Check HUMAnN3 output file and metadata input paths}
\usage{
.humann_path_checks(file_path, metadata)
}
\arguments{
\item{file_path}{Path to the HUMAnN3 output TSV file.}

\item{metadata}{Path to the metadata CSV file.}
}
\value{
'NULL' if input paths pass the checks. If any issue is found, an
error is raised.
}
\description{
This function checks that input paths of the HUMAnN3 output file and metadata
file exist and that metadata is contained in a CSV file.
}
\keyword{internal}
