% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_humann.R
\name{.humann_tbl_checks}
\alias{.humann_tbl_checks}
\title{Check validity of the tibble derived from the HUMAnN3 output file}
\usage{
.humann_tbl_checks(humann_tbl)
}
\arguments{
\item{humann_tbl}{A tibble containing the data from the HUMAnN3 output file.}
}
\value{
'NULL' if all checks pass. If any issue is found, an error is raised.
}
\description{
This function performs several checks on the tibble: its abundance values
correspond to either '# Gene Family' or '# Pathway', it is not empty, it
contains at least two samples and column classes (character/numeric) are correct.
}
\keyword{internal}
