% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/da_homogenization.R
\name{.maaslin_homogenize}
\alias{.maaslin_homogenize}
\title{Homogenize MaAsLin2 Differential Abundance Results}
\usage{
.maaslin_homogenize(tbl)
}
\arguments{
\item{tbl}{A tibble containing the MaAsLin2 DA analysis results.}
}
\value{
A tibble with a homogenized format.
}
\description{
This internal function homogenizes MaAsLin2 differential abundance (DA)
results into a consistent format with specific column names. The function
processes the input tibble and ensures that the output tibble has a
consistent structure with columns relevant for downstream analysis. It fills
certain columns with placeholder values as they are not applicable to
MaAsLin2 results.
}
