% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_lollipop.R
\name{.maaslin_lollipop}
\alias{.maaslin_lollipop}
\title{Plot MaAsLin2 Lollipop Plot}
\usage{
.maaslin_lollipop(da_result)
}
\arguments{
\item{da_result}{A tibble containing the results of the MaAsLin2 DA analysis.}
}
\value{
A ggplot2 object representing the lollipop plot.
}
\description{
This function generates a lollipop plot for MaAsLin2 differential abundance
analysis results.
}
