% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_volcano.R
\name{.maaslin_volcano}
\alias{.maaslin_volcano}
\title{Plot MaAsLin2 Volcano Plot}
\usage{
.maaslin_volcano(da_result)
}
\arguments{
\item{da_result}{A tibble containing the results of the MaAsLin2 DA analysis.}
}
\value{
A ggplot2 object representing the volcano plot.
}
\description{
This function generates a volcano plot for MaAsLin2 differential abundance
analysis results.
}
