% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_humann.R
\name{.rpk2cpm}
\alias{.rpk2cpm}
\title{Convert RPK units to CPM}
\usage{
.rpk2cpm(tbl)
}
\arguments{
\item{tbl}{A tibble containing the data from the HUMAnN3 output file.}
}
\value{
A tibble with the abundance data converted from RPK to CPM.
}
\description{
This function divides the data into two separate tibbles: one containing main
categories (gene families or pathways) and another containing the stratified
abundance values by species. After separating the data, it scales the values
within each tibble from RPK to CPM and merges normalized data into a single
tibble.
}
\keyword{internal}
