% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue_11.R
\name{draft_read_humann}
\alias{draft_read_humann}
\title{Read HUMAnN3 output file and create SummarizedExperiment objects}
\usage{
draft_read_humann(file_path, metadata)
}
\arguments{
\item{file_path}{Path to the HUMAnN3 output TSV file.}

\item{metadata}{A CSV file containing sample information.}
}
\value{
A list containing two SummarizedExperiment objects: one for groups and one for stratifications.
}
\description{
This function reads an output file from HUMAnN3 and creates SummarizedExperiment objects
containing normalized abundance values for groups (gene families or pathway categories)
and stratifications (breakdown of groups by species).
}
\examples{
file_path <-
  system.file("extdata", "All_genefam_cpm_kegg.tsv", package = "microfunk")
metadata <-
  system.file("extdata", "ex_meta.csv", package = "microfunk")
draft_read_humann(file_path, metadata)
}
