% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humann_db.R
\name{fetch_humann_db}
\alias{fetch_humann_db}
\title{Fetch HUMAnN3 databases}
\usage{
fetch_humann_db(
  annot = c("ec", "eggnog", "go", "ko", "pfam"),
  version = "v201901b",
  overwrite = FALSE
)
}
\arguments{
\item{annot}{A character vector specifying the annotations to fetch. Default
is c("ec", "eggnog", "go", "ko", "pfam").}

\item{version}{A character string representing the version of the databases
to fetch. Default is "v201901b".}

\item{overwrite}{A logical indicating whether to overwrite existing files in
the cache directory. Default is FALSE.

This function first checks if the provided annotation is valid. If
not, it aborts with a message indicating the available annotations. Then,
it creates the microfunk cache directory if it does not exist, and checks
the version compatibility. Finally, it fetches the databases from AWS and
saves them to the cache directory. If the overwrite option is TRUE,
existing files in the cache directory will be overwritten.}
}
\value{
invisible TRUE
}
\description{
This function fetches HUMAnN3 databases from AWS based on the provided
annotation and version. If the databases do not exist in the cache directory,
or if the overwrite option is enabled, the function will download the
databases from AWS.
}
\examples{
# Fetch HUMAnN3 databases with default parameters
fetch_humann_db(annot = "test")
}
