% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maaslin2.R
\name{maaslin2_quietly}
\alias{maaslin2_quietly}
\title{Run MaAsLin2 Analysis Quietly}
\usage{
maaslin2_quietly(file, metadata)
}
\arguments{
\item{file}{Path to the HUMAnN3 output (gene families or pathway) collapsed
TSV file. The file should contain aggregated abundance data across multiple
samples. Each row represents a unique functional category (gene family /
pathway) and each column corresponds to a sample.}

\item{metadata}{Path to the CSV file containing sample IDs and additional
information, where each row corresponds to a sample and each column
represents a specific annotation associated with that sample.}
}
\value{
Results of the MaAsLin2 analysis.
}
\description{
This function performs a MaAsLin2 analysis
(https://huttenhower.sph.harvard.edu/maaslin/), which finds associations
between microbiome meta-omics features and metadata. Features consist on the
abundance of gene families or metabolic pathways of multiple samples,
obtained using HUMAnN3 (https://huttenhower.sph.harvard.edu/humann). The
function takes two arguments, the path to the HUMAnN3 output file and the
metadata file, and applies linear models to asses the relationship between
microbial abundance data and metadata. It creates an output folder with a
list of significant associations in the \code{significant_results.tsv} file as
well as plots generated for each significant association, among other files.
}
\examples{
# Def data paths
metadata <-
  system.file("extdata", "ex_meta.csv", package = "microfunk")
file_path <-
  system.file("extdata", "All_genefam_cpm_kegg.tsv", package = "microfunk")

# MaAsLin2 Analysis
maaslin2_quietly(file_path, metadata)
}
