% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microfunk-package.R
\docType{package}
\name{microfunk-package}
\alias{microfunk}
\alias{microfunk-package}
\title{microfunk: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MicrobialGenomics-IrsicaixaOrg/microfunk}
  \item \url{https://microbialgenomics-irsicaixaorg.github.io/microfunk/}
  \item Report bugs at \url{https://github.com/MicrobialGenomics-IrsicaixaOrg/microfunk/issues}
}

}
\author{
\strong{Maintainer}: Francesc Català-Moll \email{fcatala@irsicaixa.es} (\href{https://orcid.org/0000-0002-2354-8648}{ORCID})

Authors:
\itemize{
  \item Judit Farré-Badia \email{judit.farre@autonoma.cat}
}

}
\keyword{internal}
