% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_abundance.R
\name{norm_abundance}
\alias{norm_abundance}
\title{Normalize gene family or pathway abundance data in a SummarizedExperiment
object}
\usage{
norm_abundance(se, norm)
}
\arguments{
\item{se}{A SummarizedExperiment object containing gene/pathway abundance
values as the assay and metadata as the column data.}

\item{norm}{A character string specifying the normalization method to apply.
Supported methods are 'cpm' for copies per million and 'relab' for relative
abundance.}
}
\value{
An updated SE object with normalized abundance values in the assay
slot.
}
\description{
This function normalizes gene family or pathway abundance data stored in a
SummarizedExperiment object, containing gene/pathway abundance values as the
assay and metadata as the column data
(https://bioconductor.org/packages/devel/bioc/vignettes/SummarizedExperiment/inst/doc/SummarizedExperiment.html),
to either CPM units (copies per million) or relative abundance.
}
\details{
It first checks whether the requested normalization method is supported. If
so, then retrieves the abundance data from the assays slot in the SE object,
determines the current normalization method applied to the data (RPK, CPM or
relative abundance) and applies the requested normalization if needed.
}
\examples{
#' # Def data paths
metadata <-
  system.file("extdata", "ex_meta.csv", package = "microfunk")
file_path <-
  system.file("extdata", "All_genefam_cpm_kegg.tsv", package = "microfunk")

# Read HUMAnN3 & Normalize to Relative Abundance
transformed_data <- read_humann(file_path, metadata) \%>\%
  norm_abundance(norm = "relab")
}
