% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_volcano.R
\name{plt_volcano}
\alias{plt_volcano}
\title{Generate a Volcano Plot for MaAsLin2 Analysis Results}
\usage{
plt_volcano(da_result)
}
\arguments{
\item{da_result}{A tibble containing the result of the differential analysis.}
}
\value{
A volcano plot visualizing differential abundance of microbial
features.
}
\description{
This function generates a volcano plot based on the results of a differential
abundance analysis in order to visualize the differential abundance of
microbial gene families or pathways (features) associated with specific fixed
effects.
}
\details{
The x-axis represents the coefficient/effect size (for MaAsLin2 DA) or
log2FoldChange (for DESeq2 DA) of each feature, indicating the magnitude and
direction of change associated with the chosen categorical variable. The
y-axis represents the statistical significance of the differences (negative
logarithm of the p-values).

Each point on the plot corresponds to a gene family or pathway, with its
position determined by both its effect size and statistical significance.
Only features with a q-value < 0.05 (for MaAsLin2) or p-adjusted value < 0.05
(for DESeq2)  are deemed significant and therefore colored, in red if its
abundance increases or blue if its abundance decreases.
}
\examples{
# Def data paths
metadata <- system.file("extdata", "ex_meta.csv", package = "microfunk")
file_path <- system.file("extdata", "All_genefam_cpm_kegg.tsv", package = "microfunk")

# Read HUMAnN3 & MaAsLin2 Analysis
da_result <-
  read_humann(file_path, metadata) \%>\%
  run_maaslin2(fixed_effects = "ARM")

# Volcano Plot
plt_volcano(da_result)
}
