% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gfam.R
\name{read_gfam}
\alias{read_gfam}
\title{Read HUMAnN Gene Families}
\usage{
read_gfam(file_path)
}
\arguments{
\item{file_path}{A string. The path to the HUMAnN gene families file.}
}
\value{
A tibble. The gene families data with columns: 'class', 'annotation', and 'norm' (normalized counts).
}
\description{
This function reads in a HUMAnN gene families file and performs several checks on the data.
It also normalizes the data and renames the columns for easier downstream analysis.
}
\examples{
file_path <-
  system.file("extdata", "demo_genefamilies.tsv", package = "microfunk")

gfam <- read_gfam(file_path)
}
