% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_read_files.R
\name{read_humann}
\alias{read_humann}
\title{Read HUMAnN data files}
\usage{
read_humann(dir, gene_pattern, path_pattern)
}
\arguments{
\item{dir}{A character string specifying the directory containing the HUMAnN data files.}

\item{gene_pattern}{A character string specifying the pattern to match gene family files.}

\item{path_pattern}{A character string specifying the pattern to match pathway abundance files.}
}
\value{
A tibble containing the merged gene family and pathway abundance data.
}
\description{
This function reads gene family and pathway abundance files from a directory.
}
