% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pathabund.R
\name{read_pathabund}
\alias{read_pathabund}
\title{read_pathabund}
\usage{
read_pathabund(path)
}
\arguments{
\item{path}{path}
}
\value{
tibble
}
\description{
read_pathabund
}
\examples{
read_pathabund(microFunk:::path_abund)
}
