% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maaslin2.R
\name{run_maaslin2}
\alias{run_maaslin2}
\title{Run MaAsLin2 Analysis Quietly}
\usage{
run_maaslin2(
  se,
  output = paste0(tempdir(), "/output_folder"),
  min_abundance = 0,
  min_prevalence = 0.1,
  min_variance = 0,
  normalization = "NONE",
  transform = "NONE",
  analysis_method = "LM",
  max_significance = 0.25,
  random_effects = NULL,
  fixed_effects = NULL,
  correction = "BH",
  standardize = TRUE,
  plot_heatmap = TRUE,
  heatmap_first_n = 50,
  plot_scatter = TRUE,
  max_pngs = 10,
  save_scatter = FALSE,
  save_models = FALSE,
  reference = NULL
)
}
\arguments{
\item{se}{A SummarizedExperiment object containing normalized abundance
values as the assay and metadata as the column data.}

\item{output}{Character string specifying the ouput directory path. Default
is a tmp folder.}

\item{min_abundance}{Minimum abundance for each feature. Default is 0.0.}

\item{min_prevalence}{Minimum percent of samples for which a feature is
detected at minimum abundance. Default is 0.1.}

\item{min_variance}{Minimum variance threshold for features to be included.
Default is 0.0.}

\item{normalization}{Normalization method to apply. Default is 'NONE'.}

\item{transform}{Transformation method to be used. Default is 'NONE'.}

\item{analysis_method}{Analysis method to apply. Default is 'LM'.}

\item{max_significance}{Q-value threshold for significance. Default is 0.25.}

\item{random_effects}{Random effects to be included in the analysis
(comma-delimited for multiple effects). Default is NULL.}

\item{fixed_effects}{Fixed effects to be included in the analysis
(comma-delimited for multiple effects). Default is NULL.}

\item{correction}{Correction method for computing the Q-value. Default is
'BH'.}

\item{standardize}{Apply Z-score so continuous metadata are on the same
scale. Default is FALSE.}

\item{plot_heatmap}{Generate a heatmap for significant associations. Default
is TRUE.}

\item{heatmap_first_n}{In the heatmap, plot only top N features with
significant associations. Default is 50.}

\item{plot_scatter}{Generate scatter plots for the significant associations.
Default is TRUE.}

\item{max_pngs}{Set the maximum number of scatter plots for significant
associations to save as png files. Default is 10.}

\item{save_scatter}{Save all scatter plot ggplot objects to an RData file.
Default is FALSE.}

\item{save_models}{Return the full model outputs and save to an RData file.
Default is FALSE.}

\item{reference}{Factor to use as reference for categorical variables,
provided as a string of 'variable,reference' (semi-colon delimited for
multiple variables).}
}
\value{
Results of the MaAsLin2 analysis on the console.
}
\description{
This function performs a MaAsLin2 analysis
(https://huttenhower.sph.harvard.edu/maaslin/), which finds associations
between microbiome meta-omics features and metadata. Features consist on the
abundance of gene families or metabolic pathways of multiple samples,
obtained using HUMAnN3 (https://huttenhower.sph.harvard.edu/humann).
}
\details{
The function takes a SummarizedExperiment object containing normalized gene/
pathway abundance values as the assay and metadata as the column data, as
well as several parameter options to run the analysis. It creates a temporary
output folder to store created files and returns the result of the analysis
on the console.
}
\examples{
# Def data paths
metadata <-
  system.file("extdata", "reduced_meta.csv", package = "microfunk")
file_path <-
  system.file("extdata", "reduced_genefam_cpm_kegg.tsv", package = "microfunk")

# Read HUMAnN3
result <- read_humann(file_path, metadata)

# MaAsLin2 Analysis
run_maaslin2(result,
             output = paste0(tempdir(), "/output_folder"),
             min_abundance = 0.0,
             min_prevalence = 0.1,
             min_variance = 0.0,
             normalization = "NONE",
             transform = "NONE",
             analysis_method = "LM",
             max_significance = 0.25,
             random_effects = NULL,
             fixed_effects = "ARM",
             correction = "BH",
             standardize = TRUE,
             plot_heatmap = TRUE,
             heatmap_first_n = 50,
             plot_scatter = TRUE,
             max_pngs = 10,
             save_scatter = FALSE,
             save_models = FALSE,
             reference = "Placebo"
              )
}
