% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomeSeq.R
\name{run_metagenomeseq}
\alias{run_metagenomeseq}
\title{Run MetagenomeSeq Analysis}
\usage{
run_metagenomeseq(
  se,
  variable,
  mod = NULL,
  coef = 2,
  B = 1,
  szero = FALSE,
  spos = FALSE
)
}
\arguments{
\item{se}{A SummarizedExperiment object containing normalized gene/pathway
abundance values.}

\item{variable}{A character string specifying the name of the variable of
interest to include in the model, matching the desired column name of the
sample metadata.}

\item{mod}{document...}

\item{coef}{document...}

\item{B}{document ..}

\item{szero}{document}

\item{spos}{document}
}
\value{
A data frame (tibble) containing the results of the MetagenomeSeq
analysis.
}
\description{
MetagenomeSeq is designed to determine features (such as Operational
Taxanomic Unit (OTU), species, genes, etc.) that are differentially abundant
between two or more groups of multiple samples
(https://www.cbcb.umd.edu/software/metagenomeSeq). It address the effects of
both normalization and under-sampling of microbial communities on disease
association detection and the testing of feature correlations. For this
function, features consist on the abundance of gene families or metabolic
pathways of multiple samples, obtained using HUMAnN3
(https://huttenhower.sph.harvard.edu/humann).
}
\details{
The function takes a SummarizedExperiment object containing normalized gene/
pathway abundance values as the assay and metadata as the column data, as
well as several parameter options to run the analysis, using a zero-inflated
Gaussian mixture model as default (\code{fitZig()}). To use a zero-inflated
Log-Normal mixture model for each feature (\code{fitFeatureModel()}), the argument
\code{model}should be modified (\code{model = "fitFeatureModel"}).It returns the
results of the analysis as a data frame (tibble).
}
\examples{
# Read HUMAnN3 & MetagenomeSeq Analysis
da_result <- read_humann(
  file_path = system.file("extdata", "All_genefam_cpm_kegg.tsv", package = "microfunk"),
  metadata = system.file("extdata", "ex_meta.csv", package = "microfunk")
 ) \%>\% run_metagenomeseq(variable = "ARM")

da_result
}
