% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_volcano.R
\name{plt_volcano}
\alias{plt_volcano}
\title{Volcano Plot}
\usage{
plt_volcano(da_result)
}
\arguments{
\item{da_result}{}
}
\value{
A volcano plot
}
\description{
Volcano Plot
}
\examples{
# Def data paths
metadata <- system.file("extdata", "ex_meta.csv", package = "microfunk")
file_path <- system.file("extdata", "All_genefam_cpm_kegg.tsv", package = "microfunk")

# Read HUMAnN3 & MaAsLin2 Analysis
da_result <-
  read_humann(file_path, metadata) \%>\%
  run_maaslin2(fixed_effects = "ARM")

# Volcano Plot
plt_volcano(da_result)
}
