% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_humann.R
\name{read_humann}
\alias{read_humann}
\title{Read HUMAnN3 output file and create SummarizedExperiment objects}
\usage{
read_humann(file_path, metadata)
}
\arguments{
\item{file_path}{Path to the HUMAnN3 output (gene families or pathway)
collapsed TSV file. The file should contain aggregated abundance data
across multiple samples. Each row represents a unique functional category
(gene family / pathway) and each column corresponds to a sample.}

\item{metadata}{Path to the CSV file containing sample IDs and additional
information, where each row corresponds to a sample and each column
represents a specific annotation associated with that sample.}
}
\value{
A SummarizedExperiment object containing normalized abundance values
as the assay and metadata as the column data.
}
\description{
This function reads a HUMAnN3 output file
(https://huttenhower.sph.harvard.edu/humann), which contains information
about the abundance of gene families or metabolic pathways of a given sample
in a TSV format. The function performs several checks on the structure and
content of the file and normalizes the data from RPK units (reads per
kilobase) to CPM units (copies per million) if needed. It also reads and
validates the metadata CSV file, ensuring it contains the required
information and matching sample IDs to the provided TSV data file. Finally,
it creates a SummarizedExperiment object
(https://bioconductor.org/packages/devel/bioc/vignettes/SummarizedExperiment/inst/doc/SummarizedExperiment.html)
containing normalized gene/ pathway abundance values as the assay and
metadata as the column data.
}
\examples{
# Def data paths
metadata <-
  system.file("extdata", "ex_meta.csv", package = "microfunk")
file_path <-
  system.file("extdata", "All_genefam_cpm_kegg.tsv", package = "microfunk")

# Read HUMAnN3
read_humann(file_path, metadata)
}
