% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_humann.R
\name{read_pathabundance}
\alias{read_pathabundance}
\title{Read HUMAnN Pathway Abundances}
\usage{
read_pathabundance(file_path)
}
\arguments{
\item{file_path}{A string. The path to the HUMAnN pathway abundance file.}
}
\value{
A tibble of the pathway abundance data with  three columns: 'class' (file type), 'annotation' (pathway)
and 'norm' (abundance units).
}
\description{
This function reads a HUMAnN pathway abundance file (https://huttenhower.sph.harvard.edu/humann),
which contains information about the abundance of metabolic pathways present in a given sample
in a .tsv format. The function performs several data validation checks, distinguishing between normalized
(CPM, copies per million)and non-normalized abundance data (RPKs, reads per kilobase).
It returns a modified tibble with renamed columns for easier downstream analysis.
}
\examples{
# Test errors
system.file("extdata", package = "microfunk") \%>\%
  list.files(pattern = "demo_pathabundance_error", full.names = TRUE) \%>\%
  purrr::walk(~ testthat::expect_error(read_pathabundance(.x)))

# Test normal file
norm_file <-
 system.file("extdata", package = "microfunk") \%>\%
 list.files(pattern = "demo_pathabundance-cpm", full.names = TRUE) \%>\%
 read_pathabundance()

testthat::expect_equal(colnames(norm_file)[3], "cpm")
file_path <-
  system.file("extdata", "demo_pathabundance.tsv", package = "microfunk")

pthab <- read_pathabundance(file_path)
}
